Added fallback for sharing
