+++ client/src/i18n.ts
@@ -86,6 +86,7 @@ const resources = {
       tasksShareBlockButtonText: 'Share tasks',
       tasksShareBlockDescription: 'Share the tasks with your friends to challenge them!',
       textForShare: 'I solved {{tasksCount}} math puzzles in @MathlyBattleBot in {{time}} seconds. Try to beat my score!',
+      textSharePointerForHandleCopy: 'Just copy this message 👇 and send it to your friend in Telegram:',
     }
   },
   ru: {
@@ -168,6 +169,7 @@ const resources = {
       tasksShareBlockButtonText: 'Поделиться примерами',
       tasksShareBlockDescription: 'Поделись примерами с друзьями, чтобы они тоже могли попробовать их решить!',
       textForShare: 'Я решил {{tasksCount}} математических примеров в @MathlyBattleBot за {{time}} секунд. Попробуй решить быстрее!',
+      textSharePointerForHandleCopy: 'Просто скопируй это сообщение 👇 и кинь другу в Телеграм:',
     }
   }
 };
